finished up the geolookup addition
there were a few things to clean up
initially, I was wanting to use the information returned from the geolookup directly, without having to do another lookup to wunderground for the city and state information.
in the end, I decided to keep that additional API call, since I would have to format the state returned from the geolookup from TX to Texas, so keeping a table of information would have been much more work, for not a lot of payout.