switched die to exit 1.
So, back when I was first learning Perl, I received some advise about exiting a program on failure.  I was specifically told, use die and not exit.  However, as I've read and researched, die will exit a program with status 255, which in the Unix world is out of range.

If someone is planning on running this script and wants the exit status for some reason, exit status 255 is most likely not going to be expected.  Exit status 0 or 1 will be expected, and are universally recognized as success and failure (from a shell or system perspective, not Perl internals).

On research, and what I'm going with for now, is that the program should exit 0 or 1 if you don't have intention of trapping the error within the script itself.  In that case, there are alternatives, such as die, or croak (for a different circumstatnce).

Anyway, my underlying point here is that sometimes people's opinion about what is right and wrong is not necessarily right or wrong.  You have to figure out through research what makes more sense for you and your application of that knowledge.  The person helping me in the beginning had good intentions, there's not doubt there, but the information which was given, IMO, wasn't correct.