removed option for zip
Issue #1

since we're doing a geolocation lookup anyway to get the other geo information, might as well remove the option to specify the location at all.
this tool is for me; I don't really expect anyone to use it other than me.  if that changes in the future, I'm happy to add the option to get weather for a specified location.